# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My API",
        "description": "Lumen Swagger API documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/students": {
            "get": {
                "summary": "Get list of students",
                "operationId": "75378cfd9c3d128e0846de3d32b101ec",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Create a new student",
                "operationId": "5e69d562cd0a6141ac0ffdd5a921d62b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nis",
                                    "name",
                                    "email",
                                    "date_of_birth",
                                    "place_of_birth",
                                    "class_name"
                                ],
                                "properties": {
                                    "nis": {
                                        "type": "string",
                                        "maxLength": 8,
                                        "example": "12345678"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Jl. Mawar No. 1"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "place_of_birth": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Jakarta"
                                    },
                                    "class_name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "XII IPA 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Student created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/students/{id}": {
            "get": {
                "summary": "Get student by ID",
                "operationId": "00eebadc92fc1451e63b944dc236ce5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update a student",
                "operationId": "ab4663f1b3d03e5867d9dce3e8fa14a3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nis",
                                    "name",
                                    "email",
                                    "date_of_birth",
                                    "place_of_birth",
                                    "class_name"
                                ],
                                "properties": {
                                    "nis": {
                                        "type": "string",
                                        "maxLength": 8,
                                        "example": "12345678"
                                    },
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "maxLength": 15,
                                        "example": "08123456789"
                                    },
                                    "address": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Jl. Mawar No. 1"
                                    },
                                    "date_of_birth": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "place_of_birth": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Jakarta"
                                    },
                                    "class_name": {
                                        "type": "string",
                                        "maxLength": 50,
                                        "example": "XII IPA 1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Student updated successfully"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete a student",
                "operationId": "ae8784bb4ca4b3b3614a7366dc0f5138",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Student deleted successfully"
                    },
                    "404": {
                        "description": "Student not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}